+++ src/main/java/com/mineastr/MineAstr.java
@@ -16,7 +16,7 @@ import org.slf4j.Logger;
 
 public final class MineAstr implements ModInitializer {
     public static final String MODID = "mineastr";
-    public static final String MOD_VERSION = "0.6.7";
+    public static final String MOD_VERSION = "0.6.9";
     public static final Logger LOGGER = LogUtils.getLogger();
 
     private static final MineAstrBridge BRIDGE = new MineAstrBridge();